// src/interfaces/FormattedAsset.ts
export default interface IAssetFormatted {
    high: string;
    low: string;
    id: string;
    last_updated: string;
    market_cap: string;
    market_cap_change: string;
    market_cap_rank: string;
    name: string;
    price: string;
    price_change: string;
    price_change_percentage: string;
    symbol: string;
    total_volume: string;
    key?: string;
}